# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Question {
  id: string;
  name: string;
  platform: string;
  tags: string[];
  approach: string;
  timeComplexity: string;
  confidence: number;
  lastRevised: string;
  mistakeNotes: string;
}