# lib/api-zod/src/generated/types/questionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuestionUpdate {
  /** @minLength 1 */
  name?: string;
  platform?: string;
  tags?: string[];
  approach?: string;
  timeComplexity?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  confidence?: number;
  lastRevised?: string;
  mistakeNotes?: string;
}